Add lombok

See merge request alexey_chirkov/universityapplication!4

+++ src/test/java/ua/com/foxminded/university/dao/impl/CourseDaoImplTest.java
@@ -25,7 +25,7 @@ public class CourseDaoImplTest {
         context = new AnnotationConfigApplicationContext(TestsContextConfiguration.class);
         courseDao = context.getBean(CourseDaoImpl.class);
         departmentDao = context.getBean(DepartmentDaoImpl.class);
-        departmentForTest = Department.builder().withId((long)0).build();
+        departmentForTest = Department.builder().withId(0L).build();
     }
 
     @Test
@@ -36,7 +36,7 @@ public class CourseDaoImplTest {
                 .build();
 
         courseDao.save(addingCourse);
-        Course readingCourse = courseDao.findById((long)5).get();
+        Course readingCourse = courseDao.findById(5L).get();
 
         assertCourses(readingCourse, addingCourse);
     }
@@ -48,7 +48,7 @@ public class CourseDaoImplTest {
                 .withDepartment(departmentForTest)
                 .build());
         courseDao.saveAll(addingCourseEntities);
-        List<Course> readingCourseEntities = Arrays.asList(courseDao.findById((long)5).get());
+        List<Course> readingCourseEntities = Arrays.asList(courseDao.findById(5L).get());
 
         assertCourses(readingCourseEntities, addingCourseEntities);
     }
@@ -56,12 +56,12 @@ public class CourseDaoImplTest {
     @Test
     void updateShouldUpdateDataOfCoursesIfArgumentIsCourses(){
         Course expected = Course.builder()
-                .withId((long)4)
+                .withId(4L)
                 .withName("Rome History")
-                .withDepartment(departmentDao.findById((long)1).get())
+                .withDepartment(departmentDao.findById(1L).get())
                 .build();
         courseDao.update(expected);
-        Course actual = courseDao.findById((long)4).get();
+        Course actual = courseDao.findById(4L).get();
 
         assertCourses(actual, expected);
     }
@@ -69,12 +69,12 @@ public class CourseDaoImplTest {
     @Test
     void updateAllShouldUpdateDataOfProfessorIfArgumentIsListOfProfessor(){
         List<Course> expected = Arrays.asList(Course.builder()
-                .withId((long)4)
+                .withId(4L)
                 .withName("Rome History")
-                .withDepartment(departmentDao.findById((long)1).get())
+                .withDepartment(departmentDao.findById(1L).get())
                 .build());
         courseDao.updateAll(expected);
-        List<Course> actual = Arrays.asList(courseDao.findById((long)4).get());
+        List<Course> actual = Arrays.asList(courseDao.findById(4L).get());
 
         assertCourses(actual, expected);
     }
@@ -82,37 +82,37 @@ public class CourseDaoImplTest {
     @Test
     void deleteShouldDeleteDataOfProfessorIfArgumentIsIdOfProfessor(){
         Optional<Course> expected = Optional.empty();
-        courseDao.deleteById((long)10);
-        Optional<Course> actual = courseDao.findById((long)10);
+        courseDao.deleteById(10L);
+        Optional<Course> actual = courseDao.findById(10L);
 
         assertThat(actual).isEqualTo(expected);
     }
 
     @Test
     void addCourseToProfessorCourseListShouldAddCourseToProfessorCourseListIfArgumentIsIdOfProfessorAndIdOfCourse(){
-        List<Course> expectedBeforeAdding = Arrays.asList (courseDao.findById((long)1).get(),
-                courseDao.findById((long)2).get());
+        List<Course> expectedBeforeAdding = Arrays.asList (courseDao.findById(1L).get(),
+                courseDao.findById(2L).get());
         List<Course> actualBeforeAdding = courseDao.findByProfessorId(7);
         assertCourses(actualBeforeAdding, expectedBeforeAdding);
 
         courseDao.addCourseToProfessorCourseList(3, 7);
 
-        List<Course> expectedAfterAdding = Arrays.asList (courseDao.findById((long)1).get(),
-                courseDao.findById((long)2).get(), courseDao.findById((long)3).get());
+        List<Course> expectedAfterAdding = Arrays.asList (courseDao.findById(1L).get(),
+                courseDao.findById(2L).get(), courseDao.findById(3L).get());
         List<Course> actualAfterAdding = courseDao.findByProfessorId(7);
         assertCourses(actualAfterAdding, expectedAfterAdding);
     }
 
     @Test
     void removeCourseToProfessorCourseListShouldRemoveCourseToProfessorCourseListIfArgumentIsIdOfProfessorAndIdOfCourse(){
-        List<Course> expectedBeforeRemoving = Arrays.asList (courseDao.findById((long)1).get(),
-                courseDao.findById((long)2).get());
+        List<Course> expectedBeforeRemoving = Arrays.asList (courseDao.findById(1L).get(),
+                courseDao.findById(2L).get());
         List<Course> actualBeforeRemoving = courseDao.findByProfessorId(7);
         assertCourses(actualBeforeRemoving, expectedBeforeRemoving);
 
         courseDao.removeCourseFromProfessorCourseList(2, 7);
 
-        List<Course> expectedAfterRemoving = Arrays.asList (courseDao.findById((long)1).get());
+        List<Course> expectedAfterRemoving = Arrays.asList (courseDao.findById(1L).get());
         List<Course> actualAfterRemoving = courseDao.findByProfessorId(7);
         assertCourses(actualAfterRemoving, expectedAfterRemoving);
     }
@@ -121,7 +121,7 @@ public class CourseDaoImplTest {
     void findByNameShouldReturnOptionalOfDepartmentEntityIfArgumentIsName(){
         Course expected = Course.builder()
                 .withName("Russia History")
-                .withDepartment(departmentDao.findById((long)1).get())
+                .withDepartment(departmentDao.findById(1L).get())
                 .build();
         Course actual = courseDao.findByName("Russia History").get();
